smart bft consensus

+++ clean.sh
@@ -2,6 +2,8 @@
 source lib/util/util.sh
 source lib.sh
 
+cleanCryptoMaterial=$1
+
 localRegistryStarted=`runningDockerContainer docker-registry`
 
 if [ -z "$localRegistryStarted" ] ; then
@@ -15,15 +17,16 @@ fi
 
 #TODO [ "${DOCKER_MACHINE_NAME}" == "orderer" ]  && EXECUTE_BY_ORDERER=1 runCLIWithComposerOverrides down || runCLIWithComposerOverrides down
 
-docker volume prune -f
-docker rmi -f $(docker images -q -f "reference=dev-*")
+if [ -z "$cleanCryptoMaterial" ]; then
+    docker volume prune -f
+    docker rmi -f $(docker images -q -f "reference=dev-*")
 
-if [ -z "$DOCKER_HOST" ] ; then
-    docker-compose -f docker-compose-clean.yaml run --rm cli.clean rm -rf crypto-config/*
-else
-    docker-machine ssh ${DOCKER_MACHINE_NAME} sudo rm -rf crypto-config
-    docker-machine ssh ${DOCKER_MACHINE_NAME} sudo rm -rf data/
+    if [ -z "$DOCKER_HOST" ]; then
+        docker-compose -f docker-compose-util.yaml run --rm cli.clean rm -rf crypto-config/*
+    else
+        docker-machine ssh ${DOCKER_MACHINE_NAME} sudo rm -rf crypto-config
+        docker-machine ssh ${DOCKER_MACHINE_NAME} sudo rm -rf data/
+    fi
 fi
-
 #docker rmi -f $(docker images -q -f "reference=olegabu/fabric-starter-client")
 #docker network rm `(docker network ls -q)`
